// gql-server/models/homeworld.model.ts
// Generated by https://quicktype.io

export interface Homeworld {
    name:            string;
    rotation_period: string;
    orbital_period:  string;
    diameter:        string;
    climate:         string;
    gravity:         string;
    terrain:         string;
    surface_water:   string;
    population:      string;
    residents:       string[];
    films:           string[];
    created:         string;
    edited:          string;
    url:             string;
}